+++ magpylib/_src/display/backend_matplotlib_auto.py
@@ -3,9 +3,9 @@ import warnings
 import matplotlib.pyplot as plt
 import numpy as np
 
-from magpylib._src.display.display_utility import MagpyMarkers
 from magpylib._src.display.traces_generic import draw_frame
 from magpylib._src.display.traces_generic import subdivide_mesh_by_facecolor
+from magpylib._src.display.traces_utility import MagpyMarkers
 
 # from magpylib._src.utility import format_obj_input
