+++ magpylib/_src/display/traces_utility.py
